refactor: remove --crash command
No need for it xd
